# docs/phase-2/openapi.yaml
openapi: 3.1.0
info:
  title: Wasselne Platform API
  version: 0.1.0
  summary: Design contract for the Wasselne rider/driver/admin API. Not an implementation.
  description: |
    Companion to `02_State_Machines_and_Transactions.md` and `schema.sql` (docs/phase-2). Money is always
    integer minor units + ISO currency code (USD, LBP). Rates/percentages are integer basis points
    (10000 = 100%). All timestamps are RFC 3339 UTC. Every mutation that could duplicate requires an
    `Idempotency-Key` header. Every response carries `X-Correlation-Id`. No example in this document is a
    real fare, phone number, legal text, or emergency number — all examples are synthetic placeholders.
servers:
  - url: https://api.wasselne.example/v1
    description: Design-time placeholder host.
security:
  - riderDriverBearer: []
...
paths:
  /auth/otp/request:
    post:
      operationId: authOtpRequest
      tags: [Auth]
      summary: "Request an SMS OTP code for sign-in or sign-up."
      security: []
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/OtpRequestBody' } } } }
      responses:
        '200': { description: "Challenge created.", headers: { X-Correlation-Id: { $ref: '#/components/headers/X-Correlation-Id' } }, content: { application/json: { schema: { $ref: '#/components/schemas/OtpRequestResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /auth/otp/verify:
    post:
      operationId: authOtpVerify
      tags: [Auth]
      summary: "Verify the OTP code. Returns a session for a known account, or a needs-profile token for a new one."
      security: []
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/OtpVerifyBody' } } } }
      responses:
        '200': { description: "OTP_INVALID / OTP_EXPIRED / OTP_RATE_LIMITED are returned as errors, not as a 200 with a flag.", headers: { X-Correlation-Id: { $ref: '#/components/headers/X-Correlation-Id' } }, content: { application/json: { schema: { $ref: '#/components/schemas/OtpVerifyResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '410': { $ref: '#/components/responses/Gone' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /me/complete-profile:
    post:
      operationId: meCompleteProfile
      tags: [Auth, Me]
      summary: "Create the profile for a new rider or driver. Bearer is the needs_profile_token from otp/verify."
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/CompleteProfileBody' } } } }
      responses:
        '201': { description: "Account created; session issued.", headers: { X-Correlation-Id: { $ref: '#/components/headers/X-Correlation-Id' } }, content: { application/json: { schema: { $ref: '#/components/schemas/SessionTokens' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }  # A-3-03: global per-IP ceiling (RATE_LIMITED)
...
  /me:
    get:
      operationId: meGet
      tags: [Me]
      summary: "Own profile."
      responses:
        '200': { description: "Profile.", headers: { X-Correlation-Id: { $ref: '#/components/headers/X-Correlation-Id' } }, content: { application/json: { schema: { $ref: '#/components/schemas/Profile' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }  # A-3-03: global per-IP ceiling (RATE_LIMITED)
    patch:
      operationId: meUpdate
      tags: [Me]
      summary: "Update language, preferred currency, name, or email."
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/ProfileUpdateBody' } } } }
      responses:
        '200': { description: "Updated profile.", headers: { X-Correlation-Id: { $ref: '#/components/headers/X-Correlation-Id' } }, content: { application/json: { schema: { $ref: '#/components/schemas/Profile' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      operationId: meDeleteAccount
      tags: [Me]
      summary: "Request account deletion, subject to retention policy."
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/DeleteAccountBody' } } } }
      responses:
        '202': { description: "Deletion accepted.", headers: { X-Correlation-Id: { $ref: '#/components/headers/X-Correlation-Id' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    riderDriverBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Rider or driver app session token. The `app` claim (RIDER|DRIVER) tags which app issued the
        token; endpoints under a given app's tags reject a token from the other app with
        NOT_AUTHORIZED.
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Admin console session token, issued only after password + TOTP MFA verification
        (POST /admin/auth/mfa/verify). Carries admin_id and role codes.
...
  headers:
    X-Correlation-Id:
      description: "Server-assigned correlation id for this request/response, also present in the error body."
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: "Request failed validation (VALIDATION_FAILED, UPLOAD_INVALID, PROMO_INVALID, ...)."
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: "Missing, expired, or invalid bearer token (NOT_AUTHENTICATED)."
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: >-
        Authenticated but not allowed (NOT_AUTHORIZED, ACCOUNT_BLOCKED, TERMS_ACCEPTANCE_REQUIRED,
        OFFER_NOT_FOR_DRIVER, CATEGORY_NOT_ELIGIBLE).
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Conflict:
      description: >-
        State conflict (VERSION_CONFLICT, ACTIVE_RIDE_EXISTS, RIDE_ALREADY_ASSIGNED,
        OFFER_ALREADY_RESOLVED, OFFER_NOT_ACKNOWLEDGED, DRIVER_NOT_ELIGIBLE, INVALID_TRIP_TRANSITION,
        IDEMPOTENCY_KEY_REUSED, QUOTE_ALREADY_USED, PAYMENT_ALREADY_RECORDED).
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Gone:
      description: "The resource expired or is no longer actionable (OFFER_EXPIRED, QUOTE_STALE, PROMO_EXPIRED)."
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: "Too many requests (OTP_RATE_LIMITED, RATE_LIMITED)."
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
        Retry-After:
          description: "Seconds until the client may retry."
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServiceUnavailable:
      description: "An upstream provider (routing, translation, storage) failed temporarily (PROVIDER_TEMPORARY_FAILURE)."
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/X-Correlation-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    ErrorCode:
      type: string
      enum:
        - VALIDATION_FAILED
        - NOT_AUTHENTICATED
        - NOT_AUTHORIZED
        - ACCOUNT_BLOCKED
        - TERMS_ACCEPTANCE_REQUIRED
        - OTP_INVALID
        - OTP_EXPIRED
        - OTP_RATE_LIMITED
        - OUTSIDE_SERVICE_AREA
        - CATEGORY_UNAVAILABLE
        - CATEGORY_NOT_ELIGIBLE
        - QUOTE_STALE
        - QUOTE_ALREADY_USED
        - ACTIVE_RIDE_EXISTS
        - PROMO_INVALID
        - PROMO_EXPIRED
        - PROMO_LIMIT_REACHED
        - DRIVER_NOT_ELIGIBLE
        - OFFER_NOT_FOR_DRIVER
        - OFFER_NOT_ACKNOWLEDGED
        - OFFER_EXPIRED
        - OFFER_ALREADY_RESOLVED
        - RIDE_ALREADY_ASSIGNED
        - INVALID_TRIP_TRANSITION
        - VERSION_CONFLICT
        - PAYMENT_METHOD_DISABLED
        - PAYMENT_ALREADY_RECORDED
        - CONTACT_WINDOW_CLOSED
        - UPLOAD_INVALID
        - IDEMPOTENCY_KEY_REUSED
        - PROVIDER_TEMPORARY_FAILURE
        - RATE_LIMITED
        - NOT_FOUND          # Phase 3 amendment A-3-02: unknown route/resource (404)
        - INTERNAL_ERROR     # Phase 3 amendment A-3-02: unhandled failure (500), no internals exposed
    Error:
      type: object
      required: [code, message, correlation_id]
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message:
          description: "Localized, user-safe message in the caller's app language."
          type: string
        correlation_id:
          type: string
          format: uuid
        details:
          description: "Optional machine-readable context (e.g. field errors). Never PII beyond what the caller already sent."
          type: object
          additionalProperties: true
    CurrencyCode:
      type: string
      enum: [USD, LBP]
...
    AppKind:
      type: string
      enum: [RIDER, DRIVER]
    Gender:
      type: string
      enum: [FEMALE, MALE]
...
    OtpRequestBody: { type: object, required: [phone_e164, app], properties: { phone_e164: { type: string, pattern: '^\+[1-9][0-9]{6,14}$' }, app: { $ref: '#/components/schemas/AppKind' } } }
    OtpRequestResponse: { type: object, required: [challenge_id, expires_at, resend_after], properties: { challenge_id: { type: string, format: uuid }, expires_at: { type: string, format: date-time }, resend_after: { type: string, format: date-time } } }
    OtpVerifyBody: { type: object, required: [challenge_id, code], properties: { challenge_id: { type: string, format: uuid }, code: { type: string, minLength: 4, maxLength: 8 } } }
    SessionTokens: { type: object, required: [access_token, refresh_token, expires_at], properties: { access_token: { type: string }, refresh_token: { type: string }, expires_at: { type: string, format: date-time } } }
    OtpVerifyResponse:
      type: object
      required: [status]
      properties:
        status:
          description: "SESSION for a returning, profiled account; NEEDS_PROFILE for a new phone number."
          type: string
          enum: [SESSION, NEEDS_PROFILE]
        session: { $ref: '#/components/schemas/SessionTokens' }
        needs_profile_token:
          description: "Short-lived token; send as bearer on POST /me/complete-profile."
          type: string
        account_id: { type: string, format: uuid }
...
    Profile:
      type: object
      required: [id, phone_e164, language, status]
      properties:
        id: { type: string, format: uuid }
        phone_e164: { type: string }
        first_name: { type: string, nullable: true }
        last_name: { type: string, nullable: true }
        gender: { $ref: '#/components/schemas/Gender' }
        gender_confirmed: { type: boolean }
        email: { type: string, nullable: true }
        language: { type: string, enum: [ar, en, fr] }
        preferred_currency: { $ref: '#/components/schemas/CurrencyCode' }
        status: { type: string, enum: [ACTIVE, DELETED] }
    CompleteProfileBody:
      type: object
      required: [first_name, last_name, gender, accepted_legal_version_ids]
      properties:
        first_name: { type: string, minLength: 1 }
        last_name: { type: string, minLength: 1 }
        gender: { $ref: '#/components/schemas/Gender' }
        email: { type: string, format: email }
        language: { type: string, enum: [ar, en, fr] }
        accepted_legal_version_ids:
          type: array
          items: { type: string, format: uuid }
    ProfileUpdateBody: { type: object, properties: { first_name: { type: string }, last_name: { type: string }, email: { type: string, format: email }, language: { type: string, enum: [ar, en, fr] }, preferred_currency: { $ref: '#/components/schemas/CurrencyCode' } } }
    DeleteAccountBody: { type: object, properties: { reason: { type: string, maxLength: 500 } } }
...